updated api validation

+++ update02/amazed/controllers/couponsController.js
@@ -1,11 +1,27 @@
 import Coupons from '../model/couponsModel.js';
+import { checkNumber } from '../utils/validator.js';
 import moment from 'moment';
 
 //Add new Coupon
 export const AddNewCoupon = async(req, res) => {
     try {
-        let date = new Date();
-        let cur_date = `${date.getDate()} ${date.toLocaleString('default', {month:'long'})} ${date.getFullYear()}`;
+
+        let cur_date = moment(new Date(Date.now())).format('YYYY-MM-DD HH:mm:ss');
+        let expiry_date = moment(new Date(req.body.expiryDate)).format('YYYY-MM-DD HH:mm:ss');
+
+        if(!req.session.user && req.session.user.role !=='Admin') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const IsValidCat = checkNumber(req.body.categoryNumber);
+        if(!IsValidCat){
+            res.send("Invalid category number")
+        }
+
+        const IsValidDiscount = checkNumber(req.body.discountPercent);
+        if(!IsValidDiscount){
+            res.send("Invalid discount")
+        }
 
         let couponData = {
             couponName: req.body.couponName,
@@ -14,7 +30,7 @@ export const AddNewCoupon = async(req, res) => {
             category: req.body.category,
             categoryNumber: req.body.categoryNumber,
             discountPercent: req.body.discountPercent,
-            expiryDate: req.body.expiryDate
+            expiryDate: expiry_date
         }
         const response = await Coupons.create(couponData)
         res.status(201).send({"success":"Successfully added the coupon"})
@@ -29,6 +45,10 @@ export const AddNewCoupon = async(req, res) => {
 export const listCoupons = async(req, res) => {
 
     try {
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const response = await Coupons.find({}).sort({createdDate: -1})
         res.status(200).send(response)
     }
@@ -41,6 +61,11 @@ export const listCoupons = async(req, res) => {
 
 // get coupon by categoryNumber
 export const listCouponsByCat = async(req, res) => {
+
+    if(!req.session.user) {
+        return res.status(400).send('No Session Found! Please Login Again')
+    }
+    
     if(req.query.category) {
         const query = req.query.category	
         try {	
@@ -67,6 +92,11 @@ export const listCouponsByCat = async(req, res) => {
 // get coupon details
 export const couponById = async(req, res) => {
     const id = req.params.id
+
+    if(!req.session.user) {
+        return res.status(400).send('No Session Found! Please Login Again')
+    }
+
     try {
         const response = await Coupons.findById(id)
         res.status(200).send(response)
@@ -82,6 +112,11 @@ export const couponById = async(req, res) => {
 export const deleteItem = async(req, res) => {
     const id = req.params.id;
     try {
+
+        if(!req.session.user && req.session.user.role !=='Admin') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const response = await Coupons.findByIdAndDelete(id);
         res.status(204).send({"success":"successfully deleted"})
     }
@@ -96,6 +131,11 @@ export const deleteItem = async(req, res) => {
 export const update = async(req, res) => {
     const id = req.params.id;
     try {
+
+        if(!req.session.user ) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const response = await Coupons.update(
             {_id: id},
             {
