+++ update02/amazed/controllers/ordersController.js
@@ -1,8 +1,43 @@
 import Orders from '../model/ordersModel.js';
+import { checkEmail, checkNumber, checkString } from '../utils/validator.js';
 
 //Place order
 export const newOrder = async(req,res) => {
+    
     try{
+
+        if(!req.session.user && req.session.user.role !=='User') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        if((req.session.user.name !== req.body.userName) || (req.session.user.email !== req.body.userEmail)) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const IsValidUserEmail = checkEmail(req.body.userEmail);
+        if(!IsValidUserEmail){
+            res.send("Invalid Email")
+        }
+
+        const IsValidEmail = checkEmail(req.body.email);
+        if(!IsValidEmail){
+            res.send("Invalid Email")
+        }
+
+        const IsValidFName = checkString(req.body.fname);
+        if(!IsValidFName){
+            res.send("Invalid first name")
+        }
+
+        const IsValidLName = checkString(req.body.lname);
+        if(!IsValidLName){
+            res.send("Invalid last name")
+        }
+
+        const IsValidPhone = checkNumber(req.body.phone);
+        if(!IsValidPhone){
+            res.send("Invalid phone number")
+        }
         let data = {
             date : req.body.date,
             userName : req.body.userName,
@@ -10,8 +45,8 @@ export const newOrder = async(req,res) => {
             orderDetails : {
                 firstName : req.body.fname,
                 lastName : req.body.lname,
-                houseAddress : req.body.houseadd,
-                apartment : req.body.apartment,
+                houseAddress : req.body.houseadd ? req.body.houseadd: '',
+                apartment : req.body.apartment ? req.body.apartment: '',
                 city : req.body.city,
                 state : req.body.state,
                 country : req.body.country,
@@ -52,10 +87,14 @@ export const newOrder = async(req,res) => {
         }
 };
 
-//get orders
+//get All orders
 export const allOrders = async(req,res) => {
     try{
-        const result = await Orders.find({})
+        if(!req.session.user && req.session.user.role !=='Admin') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const result = await Orders.find({}).sort({date: -1})
 
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -70,6 +109,10 @@ export const allOrders = async(req,res) => {
 export const ordersById = async(req,res) => {
     try{
         
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const result = await Orders.findById(req.params.id)
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -84,6 +127,11 @@ export const ordersById = async(req,res) => {
 //get orders by userEmail
 export const ordersByEmail = async(req,res) => {
     try{
+
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const userEmail = req.query.userEmail
         const result = await Orders.find({userEmail:userEmail}).sort({ date: -1})
 
@@ -101,8 +149,12 @@ export const updateOrder = async (req, res) => {
     const user = req.body;
     
     try {
-        const _id = req.params.id
 
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
+        const _id = req.params.id
         await Orders.findByIdAndUpdate(_id, {
             "delivered":req.body.delivered, 
             "orderStatus":req.body.orderStatus
@@ -120,6 +172,11 @@ export const updateOrder = async (req, res) => {
 export const deleteOrder = async (req, res) => {
 
     try {
+
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const _id = req.params.id
         const deleted = await Orders.findByIdAndDelete(_id);
