+++ update02/amazed/controllers/productsController.js
@@ -1,6 +1,7 @@
 import Products from '../model/productsModel.js';
 import Details from '../model/detailsModel.js';
 import Review from '../model/reviewModel.js';
+import moment from 'moment';
 
 //add a product
 export const addNew = async(req,res) => {
@@ -10,8 +11,11 @@ export const addNew = async(req,res) => {
     const newProduct = new Products({ asin, category, product, type, price, reviews, title, thumbnail });
     const features = req.body.features.split(',');
 
-    const date = new Date();
-    const month = date.toLocaleString('default', { month: 'long' });
+    let cur_date = moment(new Date(Date.now())).format('YYYY-MM-DD HH:mm:ss');
+
+    if(!req.session.user && req.session.user.role !=='Admin') {
+        return res.status(400).send('No Session Found! Please Login Again')
+    }
 
     let data = {
         asin: asin,
@@ -25,8 +29,8 @@ export const addNew = async(req,res) => {
         product_information: {
             dimensions: dimensions,
             weight: weight,
-            available_from: `${date.getDay()} ${month} ${date.getFullYear()}`,
-            available_from_utc: date.toISOString(),
+            available_from: cur_date,
+            available_from_utc: cur_date,
             manufacturer: manufacturer,
             model_number: model_number,
             sold_by: sold_by,
@@ -63,6 +67,11 @@ export const addNew = async(req,res) => {
 
 //add multiple products
 export const addMultiple = async(req,res) => {
+
+    if(!req.session.user && req.session.user.role !=='Admin') {
+        return res.status(400).send('No Session Found! Please Login Again')
+    }
+
     const data = req.body.products
     
     try {       
@@ -79,6 +88,11 @@ export const addMultiple = async(req,res) => {
 //get all products
 export const getProducts = async(req,res) => {
     try{
+
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         const result = await Products.find({})
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -91,6 +105,9 @@ export const getProducts = async(req,res) => {
 //get products by category or (product and type) or asin
 export const getProdsByCategory = async(req,res) => {
     try{
+        if(!req.session.user) {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
         const category = req.query.category;
         const product = req.query.product;
         const type = req.query.type;
@@ -123,6 +140,11 @@ export const getProdsByCategory = async(req,res) => {
 //delete product 
 export const deleteProduct = async (req, res) => {
     try {
+        
+        if(!req.session.user && req.session.user.role !=='Admin') {
+            return res.status(400).send('No Session Found! Please Login Again')
+        }
+
         await Products.findByIdAndDelete(req.params.id);
 
         return res.status.send({"success": "Product deleted successfully"})
